Immediate EOF detection in main loop
Put a mpsc::channel in the main loop that splits the future chain.
The "reader" stream is now read continously, so as to detect EOF.
It forwards the data onto the channel to the "writer" part.

In a next patch we'll set an atomic "invalid" flag when EOF is detected,
so that the writer part can check that in the Webnis server retry
loop, and exit early if the client has gone away.