add body_param_bytes()
Body params do not have to be valid utf-8 strings. At least not when
they are encoded using application/x-www-form-urlencoded. They can be
binary strings, or iso-8859-1, or whatever. body_param_bytes()
returns the &[u8] bytes version of the parameter.

Come to think of it, this holds true for query_param() as  well. hmm.