cache last getpwnam() name -> uid combo
If getpwuid(uid) fails, see if the last getpwnam(name) returned
that particular uid. if so, we have that name cached, and can
try a getpwnam(name) instead.

This is useful for installations where the webnis-bind
"restrict_getpwuid" setting is "true". It makes sure that getpwuid()
works at least for the last getpwnam(). Otherwise tools like `id` will
say "no such user" because the lookup by uid failed.

We might consider doing this in webnis-bind instead: cache the last 4
uids returned by a getpwnam() lookup for the requesters uid, and allow a
reverse lookup for a uid if it's in the cache for that requester.